Add command execution and exit status handling
Implemented builtin and external command distinction. Builtin commands
are implemented each in their own function are stored in array of cmd
structs which contains command's name, function pointer and
description. If command is not found in the array it is considered
external and disclamer is printed out.

Added builtin commands: debug, prompt, status, exit, help. Shell tracks
the status of the last in-foreground-executed command.

Moved parsing functionality to main.